RFID scan works fully, buzzer and light light up when a new WatCard is scanned
No need to suspend the code anymore, it now works without suspending it. PB5 pin is used to power both the light and piezo instead of before where each one uses a separate pin, making circuit design simpler. We are planning to use an audio cue to indicate when a card is rejected with 2 short flashes rather than 1 long flash